Added detection of a second use of a picture
keynote uses media references to identify a second occurance of a
picture.  We now recognizes it.

The current implementation works great if you go sequential though all
the slides but if you try to access a slide with a duplicate picture and
you haven't accessed the slide with the original it doesn't work.  I
will put in a bug on that behavior.